+++ packages/three/shaders/uniforms/hoverEffect.js
@@ -3,10 +3,21 @@ import * as THREE from "three";
 const setUniforms = (material, { map, mouse }) => {
   if (!material.uniforms.map) {
     material.uniforms.map = {};
+  }
+
+  if (!material.uniforms.mouse) {
     material.uniforms.mouse = new THREE.Vector2();
   }
-  material.uniforms.map.value = map;
-  material.uniforms.mouse.value = mouse;
+
+  if (map) {
+    const oldTexture = material.uniforms.map.value;
+    material.uniforms.map.value = map;
+    oldTexture?.dispose();
+  }
+
+  if (mouse) {
+    material.uniforms.mouse.value = mouse;
+  }
 };
 
 export default setUniforms;